Separate `StandardMapping` into `BaseMapping` and `SelectableMapping`
