+++ rdbms/src/main/java/io/spine/server/storage/jdbc/JdbcStorageFactory.java
@@ -228,7 +228,7 @@ public class JdbcStorageFactory implements StorageFactory {
         /**
          * Sets {@link TypeMapping}, which defines {@link Type} names for the used database.
          *
-         * <p>Use the {@linkplain StandardMappings#baseBuilder() base builder} to build a custom mapping.
+         * <p>Use the {@linkplain BaseMapping#baseBuilder() base builder} to build a custom mapping.
          *
          * <p>If the mapping was not specified, it is
          * {@linkplain StandardMappings#get(DataSourceWrapper) obtained} basing on
